add Contact class
