+++ src/seedu/addressbook/data/person/Contact.java
@@ -1,0 +1,24 @@
+package seedu.addressbook.data.person;
+
+public class Contact {
+    protected String value;
+    protected boolean isPrivate;
+    
+    public String toString() {
+        return value;
+    }
+
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof Address // instanceof handles nulls
+                && this.value.equals(((Address) other).value)); // state check
+    }
+
+    public int hashCode() {
+        return value.hashCode();
+    }
+
+    public boolean isPrivate() {
+        return isPrivate;
+    }
+}
